Used existing component to make accordion view in affiliation add
