+++ src/components/institution/SelectInstitutionForm.tsx
@@ -21,6 +21,7 @@ import { dataTestId } from '../../utils/dataTestIds';
 import { useDebounce } from '../../utils/hooks/useDebounce';
 import { getSortedSubUnits } from '../../utils/institutions-helpers';
 import { getLanguageString } from '../../utils/translation-helpers';
+import { OrganizationAccordion } from '../OrganizationAccordion';
 import { OrganizationRenderOption } from '../OrganizationRenderOption';
 import { OrganizationBox } from './OrganizationBox';
 
@@ -55,6 +56,8 @@ export const SelectInstitutionForm = ({
 }: SelectInstitutionFormProps) => {
   const { t } = useTranslation();
   const [searchTerm, setSearchTerm] = useState('');
+  const [selectedSubunitId, setSelectedSubunitId] = useState('');
+
   const debouncedQuery = useDebounce(searchTerm);
   const organizationSearchQuery = useSearchForOrganizations(debouncedQuery);
 
@@ -66,99 +69,103 @@ export const SelectInstitutionForm = ({
           addAffiliation(values.selectedSuggestedAffiliationId);
         } else if (values.subunit?.id) {
           addAffiliation(values.subunit.id);
+        } else if (selectedSubunitId) {
+          addAffiliation(selectedSubunitId);
         } else if (values.unit?.id) {
           addAffiliation(values.unit?.id);
         }
         setSubmitting(false);
       }}>
-      {({ isSubmitting, values, setFieldValue, resetForm }: FormikProps<OrganizationForm>) => {
-        console.log('values', values);
-        return (
-          <Form noValidate>
-            {suggestedInstitutions.length > 0 && (
-              <Paper elevation={4} sx={{ p: '1rem', maxHeight: '35vh', overflow: 'auto', mb: '1.5rem' }}>
-                <FormControl>
-                  <FormLabel>{t('registration.contributors.suggested_affiliations')}</FormLabel>
-                  <Field name={SelectOrganizationFormField.selectedSuggestedAffiliationId}>
-                    {({ field }: FieldProps<string>) => (
-                      <RadioGroup
-                        sx={{ gap: '0.5rem' }}
-                        {...field}
-                        onChange={(event) => {
-                          setSearchTerm('');
-                          resetForm();
-                          field.onChange(event);
-                        }}>
-                        {suggestedInstitutions.map((suggestedInstitution) => (
-                          <FormControlLabel
-                            sx={{
-                              '& .MuiFormControlLabel-label': {
-                                width: '100%',
-                              },
-                            }}
-                            key={suggestedInstitution}
-                            value={suggestedInstitution}
-                            control={<Radio size="small" />}
-                            label={<OrganizationBox unitUri={suggestedInstitution} />}
-                          />
-                        ))}
-                      </RadioGroup>
-                    )}
-                  </Field>
-                </FormControl>
-              </Paper>
-            )}
-            <Box sx={{ display: 'flex', flexDirection: 'column', gap: '1rem' }}>
-              <Field name={SelectOrganizationFormField.Unit}>
-                {({ field }: FieldProps<Organization>) => (
-                  <Autocomplete
-                    {...field}
-                    options={organizationSearchQuery.data?.hits ?? []}
-                    inputValue={field.value ? getLanguageString(field.value.labels) : searchTerm}
-                    getOptionLabel={(option) => getLanguageString(option.labels)}
-                    renderOption={(props, option) => (
-                      <OrganizationRenderOption key={option.id} props={props} option={option} />
-                    )}
-                    filterOptions={(options) => options}
-                    onInputChange={(_, value, reason) => {
-                      if (field.value) {
-                        setFieldValue(field.name, null);
-                      }
-                      if (reason !== 'reset') {
-                        setSearchTerm(value);
-                      }
-                    }}
-                    onChange={(_, value) => {
-                      resetForm();
-                      setFieldValue(field.name, value);
-                    }}
-                    loading={organizationSearchQuery.isPending}
-                    renderInput={(params) => (
-                      <TextField
-                        {...params}
-                        data-testid={dataTestId.organization.searchField}
-                        label={t('registration.contributors.search_for_institution')}
-                        variant="filled"
-                        fullWidth
-                      />
-                    )}
-                  />
-                )}
-              </Field>
-              {values.unit?.hasPart && values.unit.hasPart.length > 0 && (
-                <>
-                  <Typography variant="h3" component="h2" sx={{ marginTop: '1rem', fontWeight: 'normal' }}>
-                    {t('common.select_unit')}
-                  </Typography>
-                  <Field name={SelectOrganizationFormField.Subunit}>
-                    {({ field }: FieldProps<Organization>) => (
+      {({ isSubmitting, values, setFieldValue, resetForm }: FormikProps<OrganizationForm>) => (
+        <Form noValidate>
+          {suggestedInstitutions.length > 0 && (
+            <Paper elevation={4} sx={{ p: '1rem', maxHeight: '35vh', overflow: 'auto', mb: '1.5rem' }}>
+              <FormControl>
+                <FormLabel>{t('registration.contributors.suggested_affiliations')}</FormLabel>
+                <Field name={SelectOrganizationFormField.selectedSuggestedAffiliationId}>
+                  {({ field }: FieldProps<string>) => (
+                    <RadioGroup
+                      sx={{ gap: '0.5rem' }}
+                      {...field}
+                      onChange={(event) => {
+                        setSearchTerm('');
+                        resetForm();
+                        field.onChange(event);
+                      }}>
+                      {suggestedInstitutions.map((suggestedInstitution) => (
+                        <FormControlLabel
+                          sx={{
+                            '& .MuiFormControlLabel-label': {
+                              width: '100%',
+                            },
+                          }}
+                          key={suggestedInstitution}
+                          value={suggestedInstitution}
+                          control={<Radio size="small" />}
+                          label={<OrganizationBox unitUri={suggestedInstitution} />}
+                        />
+                      ))}
+                    </RadioGroup>
+                  )}
+                </Field>
+              </FormControl>
+            </Paper>
+          )}
+          <Box sx={{ display: 'flex', flexDirection: 'column', gap: '1rem' }}>
+            <Field name={SelectOrganizationFormField.Unit}>
+              {({ field }: FieldProps<Organization>) => (
+                <Autocomplete
+                  {...field}
+                  options={organizationSearchQuery.data?.hits ?? []}
+                  inputValue={field.value ? getLanguageString(field.value.labels) : searchTerm}
+                  getOptionLabel={(option) => getLanguageString(option.labels)}
+                  renderOption={(props, option) => (
+                    <OrganizationRenderOption key={option.id} props={props} option={option} />
+                  )}
+                  filterOptions={(options) => options}
+                  onInputChange={(_, value, reason) => {
+                    if (field.value) {
+                      setFieldValue(field.name, null);
+                    }
+                    if (reason !== 'reset') {
+                      setSearchTerm(value);
+                    }
+                  }}
+                  onChange={(_, value) => {
+                    resetForm();
+                    setFieldValue(field.name, value);
+                  }}
+                  loading={organizationSearchQuery.isPending}
+                  renderInput={(params) => (
+                    <TextField
+                      {...params}
+                      data-testid={dataTestId.organization.searchField}
+                      label={t('registration.contributors.search_for_institution')}
+                      variant="filled"
+                      fullWidth
+                    />
+                  )}
+                />
+              )}
+            </Field>
+            {values.unit?.hasPart && values.unit.hasPart.length > 0 && (
+              <Box sx={{ display: 'flex', flexDirection: 'column', gap: '0.5rem' }}>
+                <Typography variant="h3" component="h2" sx={{ marginTop: '1rem', fontWeight: 'normal' }}>
+                  {t('common.select_unit')}
+                </Typography>
+                <Field name={SelectOrganizationFormField.Subunit}>
+                  {({ field }: FieldProps<Organization>) => (
+                    <>
                       <Autocomplete
                         options={getSortedSubUnits(values.unit?.hasPart)}
                         getOptionLabel={(option) => getLanguageString(option.labels)}
                         renderOption={(props, option) => (
                           <OrganizationRenderOption key={option.id} props={props} option={option} />
                         )}
-                        onChange={(_, value) => setFieldValue(field.name, value)}
+                        onChange={(_, value) => {
+                          setFieldValue(field.name, value);
+                          setSelectedSubunitId(value?.id ?? '');
+                        }}
                         filterOptions={(options, state) =>
                           options.filter((option) =>
                             state
@@ -177,30 +184,38 @@ export const SelectInstitutionForm = ({
                           />
                         )}
                       />
-                    )}
-                  </Field>
-                </>
-              )}
-
-              <Box sx={{ display: 'flex', gap: '1rem', justifyContent: 'center' }}>
-                {onClose && (
-                  <Button onClick={onClose} data-testid={dataTestId.confirmDialog.cancelButton}>
-                    {t('common.cancel')}
-                  </Button>
-                )}
-                <LoadingButton
-                  variant="contained"
-                  type="submit"
-                  loading={isSubmitting}
-                  disabled={!values.unit && !values.selectedSuggestedAffiliationId}
-                  data-testid={dataTestId.registrationWizard.contributors.addSelectedAffiliationButton}>
-                  {t('common.add')}
-                </LoadingButton>
+                      {values[SelectOrganizationFormField.Unit]?.hasPart?.map((organization) => (
+                        <OrganizationAccordion
+                          key={organization.id}
+                          organization={organization}
+                          searchId={values[SelectOrganizationFormField.Subunit]?.id ?? ''}
+                          selectedId={selectedSubunitId}
+                          setSelectedId={setSelectedSubunitId}
+                        />
+                      ))}
+                    </>
+                  )}
+                </Field>
               </Box>
+            )}
+            <Box sx={{ display: 'flex', gap: '1rem', justifyContent: 'center' }}>
+              {onClose && (
+                <Button onClick={onClose} data-testid={dataTestId.confirmDialog.cancelButton}>
+                  {t('common.cancel')}
+                </Button>
+              )}
+              <LoadingButton
+                variant="contained"
+                type="submit"
+                loading={isSubmitting}
+                disabled={!values.unit && !values.selectedSuggestedAffiliationId}
+                data-testid={dataTestId.registrationWizard.contributors.addSelectedAffiliationButton}>
+                {t('common.add')}
+              </LoadingButton>
             </Box>
-          </Form>
-        );
-      }}
+          </Box>
+        </Form>
+      )}
     </Formik>
   );
 };
