+++ src/pages/search/advanced_search/OrganizationHierarchyFilter.tsx
@@ -1,10 +1,4 @@
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import RadioButtonCheckedIcon from '@mui/icons-material/RadioButtonChecked';
-import RadioButtonUncheckedIcon from '@mui/icons-material/RadioButtonUnchecked';
 import {
-  Accordion,
-  AccordionDetails,
-  AccordionSummary,
   Autocomplete,
   Box,
   Button,
@@ -21,11 +15,12 @@ import { useEffect, useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
 import { ResultParam } from '../../../api/searchApi';
+import { OrganizationAccordion } from '../../../components/OrganizationAccordion';
 import { OrganizationRenderOption } from '../../../components/OrganizationRenderOption';
 import { Organization } from '../../../types/organization.types';
 import { dataTestId } from '../../../utils/dataTestIds';
 import { getIdentifierFromId } from '../../../utils/general-helpers';
-import { getAllChildOrganizations, getSortedSubUnits } from '../../../utils/institutions-helpers';
+import { getSortedSubUnits } from '../../../utils/institutions-helpers';
 import { getLanguageString } from '../../../utils/translation-helpers';
 
 interface OrganizationHierarchyFilterProps extends Pick<DialogProps, 'open'> {
@@ -102,6 +97,8 @@ export const OrganizationHierarchyFilter = ({ organization, open, onClose }: Org
               searchId={searchId}
               selectedId={selectedId}
               setSelectedId={setSelectedId}
+              displayOrgId
+              displaySubunitsCount
             />
           ))}
         </Box>
@@ -128,99 +125,5 @@ export const OrganizationHierarchyFilter = ({ organization, open, onClose }: Org
         </Button>
       </DialogActions>
     </Dialog>
-  );
-};
-
-interface OrganizationAccordionProps {
-  organization: Organization;
-  searchId: string;
-  includeAllSubunits?: boolean;
-  level?: number;
-  selectedId: string;
-  setSelectedId: (id: string) => void;
-}
-
-const OrganizationAccordion = ({
-  organization,
-  searchId,
-  level = 0,
-  includeAllSubunits = false,
-  selectedId,
-  setSelectedId,
-}: OrganizationAccordionProps) => {
-  const { t } = useTranslation();
-
-  const isSearchedUnit = organization.id === searchId;
-  const isSelectedUnit = organization.id === selectedId;
-
-  const allSubunits = getAllChildOrganizations(organization.hasPart);
-
-  const [expandedState, setExpandedState] = useState(
-    isSelectedUnit || allSubunits.some((subunit) => subunit.id === selectedId)
-  );
-
-  if (!!searchId && !isSearchedUnit && !includeAllSubunits) {
-    if (!allSubunits.some((subunit) => subunit.id === searchId)) {
-      return null; // Hide this element if the searched ID is not a part of this unit
-    }
-  }
-
-  const expanded = expandedState || (!!searchId && !includeAllSubunits);
-  const subunitsCount = organization.hasPart?.length ?? 0;
-
-  return (
-    <Accordion
-      data-testid={dataTestId.editor.organizationAccordion(organization.id)}
-      elevation={2}
-      disableGutters
-      sx={{
-        bgcolor: level % 2 === 0 ? 'secondary.main' : 'secondary.light',
-        ml: { xs: undefined, md: level > 0 ? '1rem' : 0 },
-      }}
-      expanded={expanded}
-      onChange={() => {
-        setExpandedState(!expandedState);
-        setSelectedId(organization.id);
-      }}>
-      <AccordionSummary expandIcon={<ExpandMoreIcon sx={{ visibility: subunitsCount > 0 ? null : 'hidden' }} />}>
-        <Box
-          sx={{
-            width: '100%',
-            display: 'grid',
-            gap: '0.25rem 1rem',
-            gridTemplateColumns: {
-              xs: 'auto 1fr',
-              lg: 'auto 3fr 3fr 1fr 1fr',
-              alignItems: 'center',
-            },
-            py: '0.25rem',
-            '& > p': { fontWeight: isSearchedUnit ? 700 : undefined },
-          }}>
-          {isSelectedUnit ? <RadioButtonCheckedIcon fontSize="small" /> : <RadioButtonUncheckedIcon fontSize="small" />}
-          <Typography>{getLanguageString(organization.labels, 'nb')}</Typography>
-          <Typography sx={{ gridColumn: { xs: '1/3', lg: '3/4' } }}>{organization.labels['en']}</Typography>
-          <Typography sx={{ gridColumn: { xs: '1/3', lg: '4/5' } }}>{getIdentifierFromId(organization.id)}</Typography>
-          <Typography sx={{ gridColumn: { xs: '1/3', lg: '5/6' } }}>
-            {subunitsCount > 0 && t('editor.subunits_count', { count: subunitsCount })}
-          </Typography>
-        </Box>
-      </AccordionSummary>
-      {subunitsCount > 0 && (
-        <AccordionDetails sx={{ pr: 0 }}>
-          {expanded &&
-            organization.hasPart?.map((subunit) => (
-              <OrganizationAccordion
-                key={subunit.id}
-                organization={subunit}
-                level={level + 1}
-                searchId={searchId}
-                includeAllSubunits={includeAllSubunits || isSearchedUnit}
-                selectedId={selectedId}
-                setSelectedId={setSelectedId}
-              />
-            ))}
-        </AccordionDetails>
-      )}
-    </Accordion>
   );
 };
